# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Linguastar API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Book {
  id: number;
  title: string;
  author: string;
  description: string;
  /** @nullable */
  coverUrl?: string | null;
  price: number;
  category: string;
  language: string;
  level: string;
  isFeatured: boolean;
  isBestseller: boolean;
  totalPages: number;
  rating: number;
  reviewCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBookBody {
  title: string;
  author: string;
  description: string;
  /** @nullable */
  coverUrl?: string | null;
  price: number;
  category: string;
  language: string;
  level: string;
  isFeatured?: boolean;
  isBestseller?: boolean;
  totalPages: number;
  content: string;
}

export interface UpdateBookBody {
  title?: string;
  author?: string;
  description?: string;
  /** @nullable */
  coverUrl?: string | null;
  price?: number;
  category?: string;
  language?: string;
  level?: string;
  isFeatured?: boolean;
  isBestseller?: boolean;
  totalPages?: number;
  content?: string;
}

export interface BookContent {
  id: number;
  title: string;
  content: string;
  totalPages: number;
}

export interface Category {
  name: string;
  count: number;
  icon: string;
}

export interface Purchase {
  id: number;
  userId: string;
  bookId: number;
  amount: number;
  status: string;
  createdAt: string;
  book: Book;
}

export interface CreatePurchaseBody {
  bookId: number;
  amount: number;
}

export interface ReadingProgress {
  id: number;
  userId: string;
  bookId: number;
  currentPage: number;
  totalPages: number;
  percentComplete: number;
  lastReadAt: string;
  book: Book;
}

export interface UpdateReadingProgressBody {
  currentPage: number;
  totalPages?: number;
}

export interface UserRecord {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  role: string;
  createdAt: string;
  purchaseCount: number;
  totalSpent: number;
}

export interface UserProfile {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface RoleResponse {
  role: string;
  isAdmin: boolean;
}

export interface TopBook {
  id: number;
  title: string;
  salesCount: number;
  revenue: number;
}

export interface AnalyticsOverview {
  totalRevenue: number;
  totalUsers: number;
  totalBooks: number;
  totalPurchases: number;
  activeUsers: number;
  revenueThisMonth: number;
  newUsersThisMonth: number;
  topBooks: TopBook[];
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  bookId?: number | null;
  /** @nullable */
  amount?: number | null;
  createdAt: string;
}

export type ListBooksParams = {
  category?: string;
  search?: string;
  featured?: string;
  language?: string;
};