# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Linguastar API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List all books with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      operationId: createBook
      tags: [books]
      summary: Create a new book (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookBody"
      responses:
        "201":
          description: Book created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/{id}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get a single book by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBook
      tags: [books]
      summary: Update a book (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookBody"
      responses:
        "200":
          description: Book updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete a book (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/{id}/content:
    get:
      operationId: getBookContent
      tags: [books]
      summary: Get protected book content (must own book)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookContent"
        "403":
          description: Not purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin only)
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRecord"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [analytics]
      summary: Get analytics overview (admin only)
      responses:
        "200":
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analytics/activity:
    get:
      operationId: getRecentActivity
      tags: [analytics]
      summary: Get recent activity feed (admin only)
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        description:
          type: string
        coverUrl:
          type: ["string", "null"]
        price:
          type: number
        category:
          type: string
        language:
          type: string
        level:
          type: string
        isFeatured:
          type: boolean
        isBestseller:
          type: boolean
        totalPages:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - author
        - description
        - price
        - category
        - language
        - level
        - isFeatured
        - isBestseller
        - totalPages
        - rating
        - reviewCount
        - createdAt
        - updatedAt
    CreateBookBody:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        description:
          type: string
        coverUrl:
          type: ["string", "null"]
        price:
          type: number
        category:
          type: string
        language:
          type: string
        level:
          type: string
        isFeatured:
          type: boolean
        isBestseller:
          type: boolean
        totalPages:
          type: integer
        content:
          type: string
      required:
        - title
        - author
        - description
        - price
        - category
        - language
        - level
        - totalPages
        - content
    UpdateBookBody:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        description:
          type: string
        coverUrl:
          type: ["string", "null"]
        price:
          type: number
        category:
          type: string
        language:
          type: string
        level:
          type: string
        isFeatured:
          type: boolean
        isBestseller:
          type: boolean
        totalPages:
          type: integer
        content:
          type: string
    BookContent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        totalPages:
          type: integer
      required:
        - id
        - title
        - content
        - totalPages
...
    UserRecord:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
        createdAt:
          type: string
        purchaseCount:
          type: integer
        totalSpent:
          type: number
      required:
        - id
        - email
        - role
        - createdAt
        - purchaseCount
        - totalSpent
...
    AnalyticsOverview:
      type: object
      properties:
        totalRevenue:
          type: number
        totalUsers:
          type: integer
        totalBooks:
          type: integer
        totalPurchases:
          type: integer
        activeUsers:
          type: integer
        revenueThisMonth:
          type: number
        newUsersThisMonth:
          type: integer
        topBooks:
          type: array
          items:
            $ref: "#/components/schemas/TopBook"
      required:
        - totalRevenue
        - totalUsers
        - totalBooks
        - totalPurchases
        - activeUsers
        - revenueThisMonth
        - newUsersThisMonth
        - topBooks
    TopBook:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        salesCount:
          type: integer
        revenue:
          type: number
      required:
        - id
        - title
        - salesCount
        - revenue
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        userId:
          type: ["string", "null"]
        bookId:
          type: ["integer", "null"]
        amount:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - type
        - description
        - createdAt